director: Compiler warning fix

+++ src/director/director-request.c
@@ -48,7 +48,7 @@ director_request_get_timeout_error(struct director_request *request)
 		if (user->weak)
 			str_append(str, ", weak user");
 		str_printfa(str, ", user refreshed %u secs ago",
-			    ioloop_time - user->timestamp);
+			    (unsigned int)(ioloop_time - user->timestamp));
 	}
 	str_append_c(str, ')');
 	return str_c(str);
